Data.Generics.PlateData

Description

Deprecated: Use Data.Generics.Uniplate.Data instead

DEPRECATED: Use Data.Generics.Uniplate.Data instead.

This module exports Biplate instances for everything with Data defined. Using GHC the Data instances can be constructed with deriving Data.

Documentation

module Data.Generics.Biplate

Orphan instances

(Data a, Typeable a) => Uniplate a

Instance details

Methods

uniplate :: UniplateType a

(Data a, Data b, Uniplate b, Typeable a, Typeable b) => Biplate a b

Instance details

Methods

biplate :: BiplateType a b