Data.Generics.PlateTypeable

Description

Deprecated: Use Data.Generics.Uniplate.Typeable instead

DEPRECATED: Use Data.Generics.Uniplate.Typeable instead.

This module supplies a method for writing Biplate instances more easily.

To take an example:

data Expr = Var Int | Neg Expr | Add Expr Expr

instance Typeable Expr where ...

instance (Typeable a, Uniplate a) => PlateAll Expr a where
  plateAll (Var x  ) = plate Var |- x
  plateAll (Neg x  ) = plate Neg |+ x
  plateAll (Add x y) = plate Add |+ x |+ y

instance Uniplate Expr where
  uniplate = uniplateAll

Documentation

module Data.Generics.Biplate

module Data.Typeable

The Class

class PlateAll from to where

This class represents going from the container type to the target.

This class should only be constructed with plate, |+ and |-

Minimal complete definition

plateAll

Methods

plateAll :: from -> Type from to

uniplateAll :: PlateAll a b => a -> (Str b, Str b -> a)

This function is used to write a Uniplate instance from a PlateAll one

The Combinators

plate :: from -> Type from to

The main combinator used to start the chain.

The following rule can be used for optimisation:

plate Ctor |- x == plate (Ctor x)

(|+) :: (Typeable item, Typeable to, PlateAll item to) => Type (item -> from) to -> item -> Type from to

the field to the right may contain the target.

(|-) :: Type (item -> from) to -> item -> Type from to

The field to the right does not contain the target. This can be used as either an optimisation, or more commonly for excluding primitives such as Int.

Orphan instances

Uniplate Bool

Methods

uniplate :: UniplateType Bool

Uniplate Char

Methods

uniplate :: UniplateType Char

Uniplate Double

Methods

uniplate :: UniplateType Double

Uniplate Float

Methods

uniplate :: UniplateType Float

Uniplate Int

Methods

uniplate :: UniplateType Int

Uniplate Integer

Methods

uniplate :: UniplateType Integer

Uniplate ()

Methods

uniplate :: UniplateType ()

(Typeable * a, Typeable * b, Uniplate b, PlateAll a b) => Biplate a b

Methods

biplate :: BiplateType a b