Data.Generics.PlateData

Description

Deprecated: Use Data.Generics.Uniplate.Data instead

DEPRECATED: Use Data.Generics.Uniplate.Data instead.

This module exports Biplate instances for everything with Data defined. Using GHC the Data instances can be constructed with deriving Data.

Documentation

module Data.Generics.Biplate