Data.Generics.Str

Description

This module provides the Str data type, which is used by the underlying uniplate and biplate methods. It should not be used directly under normal circumstances.

The Data Type

data Str a

Constructors

Zero

One a

Two (Str a) (Str a)

strMap :: (a -> b) -> Str a -> Str b

strMapM :: Monad m => (a -> m b) -> Str a -> m (Str b)

strType :: Str a -> a

Take the type of the method, will crash if called

strList :: Str a -> [a]

Convert a Str to a list, assumes the value was created with listStr

listStr :: [a] -> Str a

Convert a list to a Str

strStructure :: Str a -> ([a], [a] -> Str a)

Transform a Str to a list, and back again, in a structure preserving way. The output and input lists must be equal in length.